Restructuring complete.
I've finished restructuring the directories of the doc. Each part now
resides in a directory, where the chapter directories follow. Always
include the index in the toctree - The index further contains all files
below -> Thus a clean toctree is created.

I've merged parts of the old outline to the new one were I saw fit -
however it is still messy some times and parts need to be joined/rewritten.

Not sure what to do with the rest of the outline though.

Mihi-tr