Restructuring complete...
Ok this is the second restructuring in quite a short while.

After discussions with @rgrp and Laura we decided to change the structure a
bit. The new sections are now in place and the hierarchy is a lot flatter
than before. Nevertheless there is still some room. The handbook was
stripped down to what exists at the moment.

This also introduces the "dev" branch for development of new sections etc -
please use that branch for testing and development of new content as it
will not be pushed to master imediately.

Thanks for reading!

(restructuring restructured text.. hmmm..)